+++ lib/src/gradient_button.dart
@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:gradient_widgets/src/common.dart';
 
-class GradientButton extends StatefulWidget {
+import '../gradient_widgets.dart';
+import 'gradient_widget.dart';
+
+class GradientButton extends GradientWidget {
   GradientButton(
-      {@required this.gradient,
+      {this.gradient = Gradients.hotLinear,
       @required this.child,
       @required this.callback,
+      this.shadowColor = Colors.black45,
       this.shape,
       this.shapeRadius,
       this.textStyle,
@@ -13,10 +16,31 @@ class GradientButton extends StatefulWidget {
       this.isEnabled = true,
       this.disabledGradient,
       this.increaseHeightBy = 0.0,
-      this.increaseWidthBy = 0.0});
+      this.increaseWidthBy = 0.0,
+      this.constraints,
+      this.heroTag,
+      this.tooltip,
+      this.materialTapTargetSize})
+      : super(
+            child: child,
+            callback: callback,
+            shadowColor: shadowColor,
+            shape: shape,
+            shapeRadius: shapeRadius,
+            textStyle: textStyle,
+            elevation: elevation,
+            isEnabled: isEnabled,
+            disabledGradient: disabledGradient,
+            increaseHeightBy: increaseHeightBy,
+            increaseWidthBy: increaseWidthBy,
+            constraints: constraints,
+            heroTag: heroTag,
+            tooltip: tooltip,
+            materialTapTargetSize: materialTapTargetSize);
 
   final Widget child;
   final Gradient gradient;
+  final Color shadowColor;
   final Gradient disabledGradient;
   final VoidCallback callback;
   final ShapeBorder shape;
@@ -26,108 +50,9 @@ class GradientButton extends StatefulWidget {
   final double elevation;
   final double increaseHeightBy;
   final double increaseWidthBy;
+  final BoxConstraints constraints;
+  final String tooltip;
 
-  @override
-  GradientButtonState createState() {
-    return GradientButtonState();
-  }
-}
-
-class GradientButtonState extends State<GradientButton> with SingleTickerProviderStateMixin {
-  Animation<double> _opacity;
-  AnimationController animationController;
-  bool isTappedUp = false;
-  double elevation;
-
-  Gradient gradient;
-  VoidCallback callback;
-
-  @override
-  void initState() {
-    animationController = AnimationController(duration: Duration(milliseconds: 200), vsync: this);
-
-    _opacity = Tween<double>(begin: 1.0, end: 0.8)
-        .animate(CurvedAnimation(parent: animationController, curve: Curves.fastOutSlowIn));
-
-    elevation = widget.elevation;
-    animationController.addStatusListener((status) {
-      if (animationController.isCompleted && isTappedUp) {
-        animationController.reverse();
-      }
-    });
-
-    super.initState();
-  }
-
-  void tapDown() {
-    elevation = 0.0;
-    animationController.forward();
-    isTappedUp = false;
-    setState(() {});
-  }
-
-  void tapUp() {
-    elevation = widget.elevation * 2;
-    if (!animationController.isAnimating) {
-      animationController.reverse();
-    }
-    isTappedUp = true;
-    setState(() {});
-  }
-
-
-  @override
-  void dispose() {
-    animationController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-
-    BorderRadius borderRadiusCopy = widget.shapeRadius ?? BorderRadius.circular(20.0);
-    ShapeBorder shapeCopy = widget.shape ?? RoundedRectangleBorder(borderRadius: borderRadiusCopy);
-    TextStyle textStyleCopy = widget.textStyle ?? theme.textTheme.button.copyWith(color: Colors.white);
-
-    if (widget.isEnabled) {
-      gradient = widget.gradient;
-      callback = widget.callback;
-    } else {
-      callback = null;
-      gradient = widget.disabledGradient ??
-          LinearGradient(
-            stops: widget.gradient.stops,
-            colors: const <Color>[
-              Color(0xffDADADA), // <color name="mystic">#DADADA</color>
-              Color(0xffBABEC3), // <color name="french_gray">#BABEC3</color>
-            ],
-          );
-    }
-
-    return GestureDetector(
-      onTapDown: (details) => tapDown(),
-      onTapUp: (details) => tapUp(),
-      onTapCancel: () => tapUp(),
-      child: Center(
-        child: RawMaterialButton(
-          fillColor: Colors.transparent,
-          padding: const EdgeInsets.all(0.0),
-          shape: shapeCopy,
-          elevation: elevation,
-          textStyle: textStyleCopy,
-          onPressed: () {
-            tapDown();
-            tapUp();
-            callback();
-          },
-          clipBehavior: Clip.antiAliasWithSaveLayer,
-          child: FadeTransition(
-            opacity: _opacity,
-            child: gradientContainer(context, gradient, widget.increaseHeightBy, widget.increaseWidthBy, widget.child),
-          ),
-        ),
-      ),
-    );
-  }
+  final Object heroTag;
+  final MaterialTapTargetSize materialTapTargetSize;
 }