+++ lib/src/gradient_card.dart
@@ -5,17 +5,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
+import '../gradient_widgets.dart';
+
 class GradientCard extends StatelessWidget {
-  /// Creates a material design card.
-  ///
-  /// The [clipBehavior] argument must not be null.
+
   const GradientCard({
     Key key,
-    @required this.gradient,
+    this.gradient = Gradients.hotLinear,
     this.shape,
-    this.margin = const EdgeInsets.all(4.0),
+    this.margin = const EdgeInsets.all(4),
     this.clipBehavior = Clip.none,
     this.child,
+    this.elevation,
+    this.shadowColor,
     this.semanticContainer = true,
   }) : super(key: key);
 
@@ -25,31 +27,33 @@ class GradientCard extends StatelessWidget {
   final bool semanticContainer;
   final Widget child;
   final LinearGradient gradient;
+  final double elevation;
+  final Color shadowColor;
 
   @override
   Widget build(BuildContext context) {
-//    var shape = RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0));
+//    var shape = RoundedRectangleBorder(borderRadius: BorderRadius.circular(10));
     return Semantics(
       container: semanticContainer,
       explicitChildNodes: !semanticContainer,
-      child: Container(
-        margin: margin ?? const EdgeInsets.all(4.0),
-        decoration: shape ??
-            ShapeDecoration(
-                shape: shape ??
-                    const RoundedRectangleBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(4.0)),
-                    ),
-                gradient: gradient),
-        child: Material(
-          type: MaterialType.card,
-          color: Colors.transparent,
-          elevation: 0.0,
-          shape: shape ??
-              const RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(Radius.circular(4.0)),
-              ),
-          clipBehavior: clipBehavior,
+      child: Material(
+        type: MaterialType.card,
+        color: Colors.transparent,
+        shadowColor: shadowColor ?? Colors.transparent,
+        elevation: elevation ?? 0,
+        shape: shape ??
+            const RoundedRectangleBorder(
+              borderRadius: BorderRadius.all(Radius.circular(4)),
+            ),
+        clipBehavior: clipBehavior,
+        child: Container(
+          margin: margin ?? const EdgeInsets.all(4),
+          decoration: ShapeDecoration(
+              shape: shape ??
+                  const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(4)),
+                  ),
+              gradient: gradient),
           child: child,
         ),
       ),